Uploading old iOS app project
This is an old project I worked on while learning Swift and iOS. These
are all of the files from the project with an updated README describing
what we did.

The code is written in a legacy version of Swift and needs to be
updated and corrected before it can be run. There may be further
updates to make the code compilable and possibly to add the additional
features mentioned in the proposal document